idp_stuff/integration/startup_classifier/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.7.1
black==19.3b0
bleach==3.1.0
Bottleneck==1.2.1
certifi==2019.6.16
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
dataclasses==0.6
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.9
entrypoints==0.3
fastai==1.0.51
fastprogress==0.1.21
hyperlink==17.3.1
idna==2.8
incremental==17.5.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.17
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.3
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.3.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.0
nbconvert==5.4.1
nbformat==4.4.0
nest-asyncio==1.0.0
notebook==5.7.8
numexpr==2.6.9
numpy==1.16.2
olefile==0.46
opencv-python==4.1.0.25
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
pipenv==2018.11.26
plac==0.9.6
pluggy==0.9.0
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
py==1.8.0
pyasn1==0.4.4
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
pyee==0.0.0
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyppeteer==0.0.25
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.4.1
pytest-runner==4.4
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
queuelib==1.5.0
regex==2018.1.10
requests==2.21.0
scipy==1.2.1
Scrapy==1.5.2
selenium==3.141.0
Send2Trash==1.5.0
service-identity==17.0.0
six==1.12.0
soupsieve==1.9
spacy==2.0.18
terminado==0.8.2
testpath==0.4.2
thinc==6.12.1
toml==0.10.0
toolz==0.9.0
torch==1.0.1
torchvision==0.2.2
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
Twisted==19.2.0
typed-ast==1.3.1
typing==3.6.4
ujson==1.35
urllib3==1.24.1
virtualenv==16.7.5
virtualenv-clone==0.5.3
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==6.0
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.10.11
zope.interface==4.6.0
